requirements.txt
requests
python-telegram-handler
PyYAML
web3
